requirements.txt
# Core
pytz==2022.7.1
Django==4.1.7
django-configurations==2.4
gunicorn==20.1.0
newrelic==8.7.0

# For the persistence stores
psycopg2-binary==2.9.5
dj-database-url==1.2.0

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown>=3.2
django-filter==22.1
django-cors-headers>=3.14.0
requests>=2.32

# Developer Tools
ipdb==0.13.13
ipython==8.11.0
mkdocs>=1.4.2
flake8==7.0.0
black==23.3.0

# Testing
mock==5.0.1
factory-boy==3.2.1
django-nose==1.4.7
nose-progressive==1.5.2
coverage==7.2.1
faker==18.4.0

# Static and Media Storage
django-storages==1.13.2
boto3==1.26.89

# celery
celery[redis]==5.3.1
